+++ CHANGELOG.md
@@ -6,3 +6,9 @@ The format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [0.1.0] - 2026-08-13
+
+### Added
+- Fetch 🇧🇬 holidays for the current year, clean fetched data and write to a CSV file.
+- Retry on failed fetches